requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
Flask-SQLAlchemy==2.5.1
Flask-JWT-Extended==4.3.1
Flask-Cors==3.0.10
SQLAlchemy==1.4.23
python-dotenv==0.19.0
bcrypt==3.2.0
PyJWT==2.1.0
marshmallow==3.13.0
python-dateutil==2.8.2
requests==2.26.0
gunicorn==20.1.0
python-docx==0.8.11
markdown==3.3.7
graphviz==0.20.1 